fixed tailwind bug
